+++ plug-ins/Apple.Accessibility/Apple.Accessibility_Unity/Assets/Apple.Accessibility/CHANGELOG.md
@@ -29,7 +29,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.1] - 2022-10-10
 ### Changed
-- Updated package name to `com.apple.unityplugin.accessibility`
+- Updated package name to `com.dirtybit.apple.unityplugin.accessibility`
 
 ## [1.0.0] - 2022-06-02
 ### Added