+++ plug-ins/Apple.PHASE/Apple.PHASE_Unity/Assets/CHANGELOG.md
@@ -85,7 +85,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.1] - 2022-10-10
 ### Changed
-- Updated package name to `com.apple.unityplugin.phase`.
+- Updated package name to `com.dirtybit.apple.unityplugin.phase`.
 
 ## [1.0.0] - 2022-06-02
 ### Added